requirements.txt
google-api-python-client==2.123.0
google-generativeai==0.3.1
SpeechRecognition==3.10.1
PyAudio==0.2.14  # Required for SpeechRecognition to use microphone 